Renamed categories_screen to home_navigation
